+++ src/renderer/prefComponents/keybindings/key-input-dialog.vue
@@ -19,16 +19,16 @@
             class="input-textbox"
             @keydown="handleKeyDown"
             @keyup="handleKeyUp"
-            :placeholder="placeholderText"
+            :placeholder="$t('pref.keybindings.dialog.placeholder')"
           >
         </div>
         <div class="footer">
-          <div class="descriptions">Press Enter to continue or ESC to exit.</div>
+          <div class="descriptions">{{ $t('pref.keybindings.dialog.instructions') }}</div>
           <div
             v-show="!isKeybindingValid"
             class="invalid-keybinding"
           >
-            Current key combination cannot be bound!
+            {{ $t('pref.keybindings.dialog.invalidKeybinding') }}
           </div>
         </div>
       </div>
@@ -47,10 +47,8 @@ export default {
   data () {
     this.needCommitOnClose = true
     this.currentKeybinding = null
-    this.defaultPlaceholderText = 'Press a key combination'
     return {
       showKeyInputDialog: false,
-      placeholderText: this.defaultPlaceholderText,
       isKeybindingValid: true,
       keybindingInputValue: ''
     }